+++ pytsite/formatters/__init__.py
@@ -5,4 +5,4 @@ __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
 from ._formatters import Formatter, Transform, Bool, Int, PositiveInt, AboveZeroInt, Float, PositiveFloat, Str, Enum, \
-    DateTime, JSON, JSONArrayToList, JSONArrayToTuple, JSONObjectToDict
+    DateTime, JSON, JSONArray, JSONObject