+++ pytsite/http/__init__.py
@@ -6,6 +6,7 @@ __license__ = 'MIT'
 
 # Public API
 from . import _request as request, _response as response, _session as session, _error as error
+from ._headers import Headers
 from ._request import Request
 from ._response import Response, Redirect as RedirectResponse, JSON as JSONResponse
 from ._session import Session