+++ pytsite/http/_response.py
@@ -1,11 +1,13 @@
 """PytSite Response Objects
 """
-import json
-from werkzeug.wrappers import Response as _BaseResponse
-
 __author__ = 'Oleksandr Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
+
+
+import json
+from werkzeug.wrappers import Response as _BaseResponse
+from ._headers import Headers as _Headers
 
 
 class Response(_BaseResponse):
@@ -29,7 +31,7 @@ class JSON(Response):
     """JSON HTTP response.
     """
 
-    def __init__(self, content, status: int = 200):
+    def __init__(self, content, status: int = 200, headers: _Headers = None):
         """Init.
         """
-        super().__init__(json.dumps(content), content_type='application/json', status=status)
+        super().__init__(json.dumps(content), status, headers, content_type='application/json')